Edit users is working

+++ AssetManager.Application/Users/Dto/UserDto.cs
@@ -15,7 +15,7 @@ namespace AssetManager.Users.Dto
         public string EmailAddress { get; set; }
         public DateTime? LastLoginTime { get; set; }
         public bool IsActive { get; set; }
-        public long? TenantId { get; set; }
+        public int? TenantId { get; set; }
         public string TenancyName { get; set; }
     }
 }
